requirements.txt
dewloosh.core>=1.0.14
numba
numpy
scipy
awkward==1.10.2
sympy
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
wheel>=0.38.0 # not directly required, pinned by Snyk to avoid a vulnerability